backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-filter==23.2
djoser==2.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==0.21.0
Pillow==10.1.0
drf-extra-fields==3.4.0
django-cors-headers==4.3.1 